Move ValuedEvent to chia.types.transaction_queue_entry
