+++ chia/_tests/util/test_misc.py
@@ -11,10 +11,11 @@ from chia._tests.util.misc import RecordingWebServer
 from chia._tests.util.split_managers import SplitAsyncManager, SplitManager, split_async_manager, split_manager
 from chia.cmds.cmds_util import format_bytes, format_minutes, validate_directory_writable
 from chia.types.blockchain_format.program import Program
+from chia.types.transaction_queue_entry import ValuedEvent
 from chia.util.batches import to_batches
 from chia.util.errors import InvalidPathError
 from chia.util.ints import uint64
-from chia.util.misc import ValuedEvent, satisfies_hint
+from chia.util.misc import satisfies_hint
 from chia.util.timing import adjusted_timeout, backoff_times
 
 T = TypeVar("T")
